requirements.txt
absl-py==0.2.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.6.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==1.5.0
blinker==1.4
cachetools==4.2.1
Cerberus==1.2
certifi==2018.4.16
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==5.0.5
defusedxml==0.7.1
entrypoints==0.3
fastapi==0.61.2
Flask==1.0.2
gast==0.2.0
gitdb==4.0.7
GitPython==3.1.14
h11==0.11.0
html5lib==0.9999999
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==0.24
jedi==0.18.0
Jinja2==2.10
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.0
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.4
opencv-python==4.4.0.46
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.15.7
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pydantic==1.7.2
pydeck==0.6.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzbar==0.1.8
pyzmq==22.0.3
requests==2.25.1
Send2Trash==1.5.0
six==1.11.0
smmap==4.0.0
starlette==0.13.6
streamlit==0.79.0
terminado==0.9.4
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
torch
torchvision
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
uvicorn==0.12.2
validators==0.18.2
watchdog==2.0.2
wcwidth==0.2.5
Werkzeug==0.14.1
widgetsnbextension==3.5.1
